// src/openapi/generated/models/measurement-daily-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Esquema para el resumen diario de mediciones por residente.
 *
 * Attributes:
 *     resident_id (str): ID del residente
 *     resident_full_name (str): Nombre completo del residente
 *     bed_name (Optional[str]): Nombre de la cama asignada
 *     date (date): Fecha de las mediciones
 *     measurement_count (int): Número total de mediciones del día
 *     measurement_types (List[str]): Tipos de mediciones realizadas
 *     first_measurement_time (str): Hora de la primera medición (HH:MM:SS)
 *     last_measurement_time (str): Hora de la última medición (HH:MM:SS)
 */
export interface MeasurementDailySummary {
  bed_name?: (string | null);
  date: string;
  first_measurement_time: string;
  last_measurement_time: string;
  measurement_count: number;
  measurement_types: Array<string>;
  resident_full_name: string;
  resident_id: string;
}