// src/openapi/generated/models/paginated-response-measurement-daily-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MeasurementDailySummary } from '../models/measurement-daily-summary';
export interface PaginatedResponseMeasurementDailySummary {
  has_next: boolean;
  has_prev: boolean;
  items: Array<MeasurementDailySummary>;
  page: number;
  pages: number;
  size: number;
  total: number;
}